+++ src/cadqueryhelper/shape/arch.py
@@ -38,3 +38,15 @@ def arch_pointed(length=30, width=5, height=50, inner_height=25):
     result.metadata = meta
 
     return result
+
+def arch_round(length=30, width=5, height=50):
+    #create the initial shape
+    arch = (cq.Workplane("XY")
+              .box(length, width, height)
+              .translate((0,0,(height/2)))
+              )
+
+    # round off the top
+    arch = arch.faces("Z").edges("Y").fillet((length/2)-.01)
+
+    return arch