+++ src/cadqueryhelper/wave/__init__.py
@@ -1,2 +1,3 @@
 from .triangle import triangle
 from .sawtooth import sawtooth
+from .square import square