+++ src/cadqueryhelper/wave/__init__.py
@@ -1,3 +1,4 @@
 from .triangle import triangle
 from .sawtooth import sawtooth
 from .square import square
+from .sine import sine