+++ changes.md
@@ -1,6 +1,7 @@
 # cadqueryhelper Changelog
 
 ## main wip
+* Added arch round
 
 ## 0.0.8
 * Updated cadquery version